Correction to the loading of an intialization file.
So, if the file is specified on the command line, with -init, then we
don't apply the default file spec.  However, if -init is not specified, the
file SQLITE_INIT, with a default spec of SYS$LOGIN:.DAT is opened.

This corrects behaviour noticed by @hb-- and reported in the comments of
issue #5.